[Move] delete 테스트 코드 옮김
